1. create vehicle network monitor project; 2. fixed delay_analysis bug
