+++ eBPF_Documentation/content/projects/01_Vehicle_network_monitoring/01_meeting/2023-02-16.md
@@ -1,0 +1,55 @@
+# 上汽网络小分队会议纪要-2023-02-16
+
+时间：2023年2月16日
+
+参会人：张毅峰、程嘉辉、付紫杨、白宇宣
+
+---
+
+讨论事项：
+
+### 1. 开发项目
+
+#### 1.1 备选项目
+
+##### 项目1：车载网络性能观测及网络抖动预警（第一优先级）
+
+1. 网络数据包的调用栈回溯（参考nettrace）
+2. 协议栈耗时统计（高细粒度版 delay_analysis）
+3. 网络状态监控及网络抖动预警（可部分参考nettrace）
+4. 【可选】TSN及非TSN网络场景下，我们工具的性能表现（注重可靠性）
+5. 性能测试/压力测试
+6. 数据呈现形式（先命令行形式呈现）
+
+##### 项目2：虚拟化+容器化多操作系统多应用场景网络通信优化
+
+1. 减少CPU负载，减少数据包收发路径
+
+##### 项目3：手机监控项目
+
+##### 项目4：自动驾驶场景在线数据过滤
+
+1. 是否可以在不影响自动驾驶算法运行的前提下，无感获取自动驾驶算法优化感兴趣的数据。由内核进行数据过滤。
+
+#### 1.2 技术路线
+
+1. 先在X64平台运行，后续移植至 ARM64
+2. 基于libbpf
+
+#### 1.3 任务安排
+
+1. 理解nettrace架构
+2. 移植delay_analysis至libbpf
+
+### 2. 本周五汇报
+
+---
+
+会议决议：
+
+1. 以网络性能观测及网络抖动预警为第一优先级
+2. 任务安排方面，目前还只涉及基础知识，因此不分工，各自实现，提高自己的eBPF编程技能
+3. 本周五汇报由付紫杨负责
+4. 下周有进展，周四（2月23日）下午一点半交流，若无阶段性进展，下下周（3月02日）交流
+
+---
