+++ eBPF_Documentation/content/projects/01_Vehicle_network_monitoring/02_work_schedule/2023-02-20_24/2023-02-20_24.md
@@ -1,0 +1,12 @@
+# 2023-02-20_24
+
+### 1. 本周工作
+1. 确定项目，先做起来，后续按需更改
+
+2. 移植bcc中的delay_analysis至libbpf_bootstrap上（各自实现）
+
+   ![image-20230223105800611](/home/fzy/Downloads/04_bcc_ebpf/pr_lmp/lmp/eBPF_Documentation/content/projects/network/02_Works/2023-02-20_24/2023-02-20_24.assets/image-20230223105800611.png)
+### 2. 下一步计划：
+1. 尝试在Orin开发板运行delay_analysis
+2. 理解 nettrace 原理
+
